Remove hard coded stat values - the stat bonuses (as well as max stat) can be set in a file that is read at load time.  The default lib/stat_bonus file has bonuses exactly the same as the old hard coded values (or in case of Charisma, as close as could be done with the change).  Note: you must do a 'make install' in the lib directory after this update, otherwise, your server will be missing the stat bonuses and fail to run. Automake/autoconf should automatically be run to know about the stat_bonus file. MSW 2011-05-01
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@14447 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/libproto.h
@@ -148,7 +148,7 @@ extern void change_exp(object *op, sint64 exp, const char *skill_name, int flag)
 extern void apply_death_exp_penalty(object *op);
 extern int did_make_save(const object *op, int level, int bonus);
 extern void share_exp(object *op, sint64 exp, const char *skill, int flag);
-extern float get_cha_bonus(int stat);
+extern int get_cha_bonus(int stat);
 extern int get_dex_bonus(int stat);
 extern int get_thaco_bonus(int stat);
 extern uint32 get_weight_limit(int stat);
